Fix paralleling efficiency gain: float to double

+++ lib/headers/generator.h
@@ -1,3 +1,3 @@
 #define SIZE 500000000
 
-void gen_vector(float *);
+void gen_vector(double *);
